swapped ints to float for more precision

+++ AdvancedGameTechnologies/CSC8503/NavMeshNode.h
@@ -7,7 +7,7 @@ namespace NCL {
 
 		struct Edge {
 			NavMeshNode* neighbour;
-			int cost;
+			float cost;
 		};
 
 		class NavMeshNode {
@@ -26,7 +26,7 @@ namespace NCL {
 			}
 			bool DetectImpassableEdge(NavMeshNode* neighbour);
 			void CalculateEdges();
-			void AddEdge(NavMeshNode* neighbour, int cost);
+			void AddEdge(NavMeshNode* neighbour, float cost);
 			std::vector<Edge> GetEdges() { return this->edges; }
 			NCL::Maths::Vector3 GetPosition() {	return this->position; }
 			float GetGScore(){ return this->gScore; }
